backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic==2.5.0
pydantic-settings==2.1.0

# Payment Integration
stripe==7.5.0
paypalrestsdk==1.13.3

# File Storage & Processing
boto3==1.34.0
python-magic==0.4.27
pillow==10.1.0

# Email
fastapi-mail==1.4.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
celery==5.3.4
flower==2.0.1

# AI & Code Analysis
openai==1.3.7
pygments==2.17.2
radon==6.0.1
bandit==1.7.5

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# CORS
fastapi-cors==0.0.6
